refactor: Add inner search sidebar component
